fix: tighten library API surface
